Create README.md
Level 1:
● Train a cigarette detector model (select architecture,
find/prepare data, train). It is not allowed to use test data in
training. The training cycle should be written
independently in Pytorch/Tensorflow. The use of ready-made platforms
for learning “without code" is not allowed.
● Implement an inference on the image from the command line (draw
bboxes).
Startup example: python3 detector.py --source 00001.png
As a result, the file 00001_out.png should appear in the current folder, on
which the bounding boxes are drawn.
Level 2:
● Implement a detector predicate with real-time rendering (in a separate
stream) on video recordings.
Startup example: python3 detector.py --source video.mp4
As a result, a video with rendered bounding boxes should be played.
At the same time, the video playback should not slow down, and should go at the
same frame rate as the original video.